chore: Start removing RotateWithDirection

+++ crates/buttplug_server/src/device/protocol_impl/joyhub/joyhub_v6.rs
@@ -65,9 +65,9 @@ impl ProtocolHandler for JoyHubV6 {
     &self,
     feature_index: u32,
     _feature_id: Uuid,
-    speed: u32,
+    speed: i32,
   ) -> Result<Vec<HardwareCommand>, ButtplugDeviceError> {
-    self.form_hardware_command(feature_index, speed)
+    self.form_hardware_command(feature_index, speed as u32)
   }
 
   fn handle_output_oscillate_cmd(
